fix(metal): drop a no-op f32 cast in the gemma3 attention parity test
`cargo clippy --all-targets -- -D warnings` fails on rustc 1.94.1 with

    error: casting to the same type is unnecessary (`f32` -> `f32`)
      --> src/metal.rs:46847:34

`scores` is a `Vec<f32>` and `m`/`l` are both `f32`, so `(scores[p2] - m).exp() / l`
is already `f32` and the trailing `as f32` does nothing. Drop it, and the redundant
parentheses with it.

The lint sits in `mod attn_mm_parity`, which is gated on
`cfg(all(test, target_os = "macos"))`, so only the macOS `--all-targets` leg
compiles it. It has been on main since ee76a22f (2026-07-31) without CI noticing,
because CI pins an older toolchain than the one in use locally:
`.github/workflows/ci.yml` selects `dtolnay/rust-toolchain@1.89.0` while
`rust-toolchain.toml` asks for 1.95.0. Reconciling those pins is left as separate
work; this change is a no-op either way.

Verified on macOS/arm64 at rustc 1.94.1: `cargo fmt --all -- --check` and
`cargo clippy --all-targets -- -D warnings` both clean (reverting the one line
reproduces exactly this error and nothing else), and `cargo test --lib` is
2338 passed / 24 ignored, with two pre-existing `fabric::http` TLS failures that
reproduce unchanged on main.

+++ src/metal.rs
@@ -46844,7 +46844,7 @@ mod attn_mm_parity {
                 for d in (0..hd).step_by(3) {
                     let mut o = 0f32;
                     for p2 in 0..=q {
-                        let pw = ((scores[p2] - m).exp() / l) as f32;
+                        let pw = (scores[p2] - m).exp() / l;
                         // kernel rounds P to half
                         let pw = fh(h(pw));
                         o += pw * fh(v[(z / group) * max_pos * hd + p2 * hd + d]);
